Correcciones del apartado de medidas y se reaplicaron los cambios en los archivos correspondientes.

+++ UsaloYa.Dto/SetStockDto.cs
@@ -3,7 +3,7 @@
     public class SetStockDto
     {
         public int ProductId { get; set; }
-        public int UnitsInStock { get; set; }
+        public decimal UnitsInStock { get; set; }
         public bool IsHardReset { get; set; }
     }
 }
